Added Api Endpoints and Repository update

+++ Movies.Backend/Repository/MovieRepository.cs
@@ -1,5 +1,8 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using AutoMapper;
+using Microsoft.CodeAnalysis;
+using Microsoft.EntityFrameworkCore;
 using Movies.Backend.Models;
+using Movies.Backend.Models.DataTransferObjects;
 using Movies.Backend.Repository.IRepository;
 using Movies.Data;
 using System.Collections.Generic;
@@ -10,36 +13,66 @@ namespace Movies.Backend.Repository
     public class MovieRepository : IMovieRepository
     {
         private readonly ApplicationDbContext _db;
+        private IMapper _mapper;
 
-        public MovieRepository(ApplicationDbContext db)
+        public MovieRepository(ApplicationDbContext db, IMapper mapper)
         {
             _db = db;
+            _mapper = mapper;
         }
 
-        public void AddMovie(Movie movie)
+        public async Task<MovieDto> AddMovie(MovieDto movieDto)
         {
+            Movie movie = _mapper.Map<MovieDto, Movie>(movieDto);
             _db.Movies.Add(movie);
+            await _db.SaveChangesAsync();
+
+            return _mapper.Map<Movie, MovieDto>(movie);
         }
 
-        public Movie? GetMovie(int? id)
+
+        public async Task<MovieDto> UpdateMovie(MovieDto movieDto)
         {
-            return _db.Movies.FirstOrDefault(m => m.Id == id);
+            Movie movie = _mapper.Map<MovieDto, Movie>(movieDto);
+            _db.Movies.Update(movie);
+            await _db.SaveChangesAsync();
+
+            return _mapper.Map<Movie, MovieDto>(movie);
         }
 
-        public IEnumerable<Movie> GetAllMovies(Expression<Func<Movie, bool>>? filter = null)
+        public async Task<MovieDto> GetMovie(int? movieId)
         {
-            List<Movie> movies = _db.Movies.Where(filter).ToList(); 
-            return movies;
+            Movie movie = await _db.Movies.FirstOrDefaultAsync(m => m.Id == movieId);
+            return _mapper.Map<MovieDto>(movie);
         }
 
-        public void RemoveMovie(Movie movie)
+        public async Task<IEnumerable<MovieDto>> GetAllMovies()
         {
-            _db.Movies.Remove(movie);
+            List <Movie> movies =await _db.Movies.ToListAsync(); 
+            return _mapper.Map<List<MovieDto>>(movies);
         }
 
-        public void Save()
+        public async Task<bool> RemoveMovie(int movieId)
         {
-            _db.SaveChanges();
+            try
+            {
+                Movie movie = await _db.Movies.FirstOrDefaultAsync(u => u.Id == movieId);
+
+                if (movie == null)
+                {
+                    return false;
+                }
+
+                _db.Movies.Remove(movie);
+                await _db.SaveChangesAsync();
+                return true;
+            }
+            catch (Exception)
+            {
+                return false;
+            }
+
         }
+
     }
 }
